feat: batch bet creation with random amounts + raise open bet limit to 255

+++ contracts/coinflip-pvp-vault/src/error.rs
@@ -22,7 +22,7 @@ pub enum ContractError {
     },
 
     #[error("Too many open bets: max {max}")]
-    TooManyOpenBets { max: u8 },
+    TooManyOpenBets { max: u16 },
 
     #[error("Bet amount below minimum: {min}")]
     BetAmountBelowMinimum { min: String },
