+++ contracts/coinflip-pvp-vault/src/msg.rs
@@ -9,7 +9,7 @@ pub struct InstantiateMsg {
     pub commission_bps: u16,
     pub min_bet: Uint128,
     pub reveal_timeout_secs: u64,
-    pub max_open_per_user: u8,
+    pub max_open_per_user: u16,
     pub max_daily_amount_per_user: Uint128,
     /// Open bet TTL in seconds (0 = no expiry). Default: 43200 (12h).
     #[serde(default = "crate::state::default_bet_ttl_secs")]
@@ -55,7 +55,7 @@ pub enum ExecuteMsg {
         commission_bps: Option<u16>,
         min_bet: Option<Uint128>,
         reveal_timeout_secs: Option<u64>,
-        max_open_per_user: Option<u8>,
+        max_open_per_user: Option<u16>,
         max_daily_amount_per_user: Option<Uint128>,
         bet_ttl_secs: Option<u64>,
     },
@@ -118,7 +118,7 @@ pub struct ConfigResponse {
     pub commission_bps: u16,
     pub min_bet: Uint128,
     pub reveal_timeout_secs: u64,
-    pub max_open_per_user: u8,
+    pub max_open_per_user: u16,
     pub max_daily_amount_per_user: Uint128,
     pub bet_ttl_secs: u64,
 }
